refactored tests into functional and unit, and added a test for testing that function maintains its original functionality
